Separating webhook types into their own file
