+++ src/test/context.test.ts
@@ -4,7 +4,7 @@ import Cookies from 'cookies';
 
 import * as ShopifyErrors from '../error';
 import {Context} from '../context';
-import {ApiVersion, ContextParams} from '../types';
+import {ApiVersion, ContextParams} from '../base_types';
 
 jest.mock('cookies');
